containerd build clean on Solaris
