+++ ctr/container_solaris.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"errors"
+)
+
+func createStdio() (s stdio, err error) {
+	return s, errors.New("createStdio not implemented on Solaris")
+}
