+++ specs/spec_solaris.go
@@ -1,0 +1,8 @@
+package specs
+
+import ocs "github.com/opencontainers/specs/specs-go"
+
+type (
+	ProcessSpec ocs.Process
+	Spec        ocs.Spec
+)
